Move BossAI to its own header
