+++ src/game/AI/ScriptDevAI/base/CombatAI.cpp
@@ -93,44 +93,3 @@ void CombatAI::KilledUnit(Unit* victim)
         ResetTimer(ACTION_ON_KILL_COOLDOWN, 10000);
     }
 }
-
-void BossAI::AddOnDeathText(uint32 text)
-{
-    m_onKilledTexts.push_back(text);
-}
-
-void BossAI::AddOnAggroText(uint32 text)
-{
-    m_onAggroTexts.push_back(text);
-}
-
-void BossAI::JustDied(Unit* killer)
-{
-    if (!m_onKilledTexts.empty())
-        DoBroadcastText(m_onKilledTexts[urand(0, m_onAggroTexts.size() - 1)], m_creature, killer);
-    if (m_instanceDataType == -1)
-        return;
-    if (ScriptedInstance* instance = static_cast<ScriptedInstance*>(m_creature->GetInstanceData()))
-        instance->SetData(m_instanceDataType, DONE);
-    CombatAI::JustDied(killer);
-}
-
-void BossAI::JustReachedHome()
-{
-    if (m_instanceDataType == -1)
-        return;
-    if (ScriptedInstance* instance = static_cast<ScriptedInstance*>(m_creature->GetInstanceData()))
-        instance->SetData(m_instanceDataType, FAIL);
-}
-
-void BossAI::Aggro(Unit* who)
-{
-    m_combatStartTimestamp = std::chrono::steady_clock::now();
-
-    if (!m_onAggroTexts.empty())
-        DoBroadcastText(m_onAggroTexts[urand(0, m_onAggroTexts.size() - 1)], m_creature, who);
-    if (m_instanceDataType == -1)
-        return;
-    if (ScriptedInstance* instance = static_cast<ScriptedInstance*>(m_creature->GetInstanceData()))
-        instance->SetData(m_instanceDataType, IN_PROGRESS);
-}
