+++ src/game/AI/ScriptDevAI/base/CombatAI.h
@@ -55,53 +55,4 @@ class CombatAI : public ScriptedAI
         bool m_stopTargeting;
 };
 
-class BossAI : public CombatAI
-{
-    public:
-        BossAI(Creature* creature, uint32 combatActions) : CombatAI(creature, combatActions)
-        {}
-        /**
-        * Adds one or more Broadcast Texts to possibly emit when Unit dies
-        * This function is not called if JustDied is overridden. Add CombatAI::JustDied(); to your overriding function.
-        * @param text Broadcast Text ID
-        */
-        void AddOnDeathText(uint32 text);
-        template<typename... Targs>
-        void AddOnDeathText(uint32 value, Targs... fargs)
-        {
-            AddOnDeathText(value);
-            AddOnDeathText(fargs...);
-        }
-        /**
-        * Adds one or more Broadcast Texts to possibly emit when Unit enters combat
-        * This function is not called if Aggro is overridden. Add CombatAI::Aggro(); to your overriding function.
-        * @param text Broadcast Text ID
-        */
-        void AddOnAggroText(uint32 text);
-        template<typename... Targs>
-        void AddOnAggroText(uint32 value, Targs... fargs)
-        {
-            AddOnAggroText(value);
-            AddOnAggroText(fargs...);
-        }
-
-        void SetDataType(uint32 type) { m_instanceDataType = type; }
-
-        void JustDied(Unit* killer = nullptr) override;
-        void JustReachedHome() override;
-        void Aggro(Unit* who = nullptr) override;
-
-        std::chrono::seconds TimeSinceEncounterStart()
-        {
-            return std::chrono::duration_cast<std::chrono::seconds>(std::chrono::steady_clock::now() - m_combatStartTimestamp);
-        }
-    private:
-        std::vector<uint32> m_onKilledTexts;
-        std::vector<uint32> m_onAggroTexts;
-
-        uint32 m_instanceDataType = -1;
-
-        std::chrono::steady_clock::time_point m_combatStartTimestamp;
-};
-
 #endif
